requirements.txt
discord.py==2.4.0
aiohttp==3.9.1
python-dotenv==1.0.0
aiosqlite==0.19.0
